+++ src/components/ErrorBoundary/_errorBoundary.scss
