+++ src/components/MainPage/_mainPage.scss
