+++ src/components/NotFound/404.test.tsx
@@ -1,6 +1,6 @@
 import { describe, expect, it } from 'vitest';
 import { render, screen } from '@testing-library/react';
-import App from '../App_/App';
+import App from '../App/App';
 import { MemoryRouter } from 'react-router-dom';
 
 describe('Tests for the 404 Page component:', () => {
