+++ src/components/Pokemon/PokemonList/PokemonList.test.tsx
@@ -1,7 +1,7 @@
 import { describe, expect, it } from 'vitest';
 import { render, screen } from '@testing-library/react';
 import { BrowserRouter } from 'react-router-dom';
-import App from '../../App_/App';
+import App from '../../App/App';
 import { server } from '../../../mocks/server';
 import { HttpResponse, http } from 'msw';
