+++ src/components/Search/SearchInput.tsx
@@ -1,6 +1,6 @@
 import { useContext } from 'react';
 
-import { PokemonContext } from '../Context_/PokemonContextProvider';
+import { PokemonContext } from '../Context/PokemonContextProvider';
 
 export const SearchInput = ({
   searchPokemon,
